requirements.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
black==25.9.0
celery==5.5.3
click==8.3.0
Django==5.2.7
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
factory-boy==3.3.3
Faker==37.12.0
flake8==7.3.0
iniconfig==2.3.0
isort==7.0.0
kombu==5.5.4
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pytest-xdist==3.6.1
python-dotenv==1.1.1
pytokens==0.2.0
redis==7.0.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
django-debug-toolbar==6.0.0
